feat: retake image with max attempts
