+++ Sources/Alloy/Internal/Views/Scan Instructions/ScanInstructionsView.swift
@@ -122,120 +122,6 @@ internal struct ScanInstructionsView: View {
     
 }
 
-internal struct ScanFooter: View {
-    
-    // MARK: - Properties
-    var documentType: DocumentType?
-    var retry: () -> Void
-
-    @EnvironmentObject private var configViewModel: ConfigViewModel
-    @State private var showNext = false
-    
-    // MARK: - Main
-    var body: some View {
-        
-        VStack(spacing: 20) {
-            
-            Button {
-                
-                configViewModel.markCurrentStepCompleted(documentSelected: documentType)
-                showNext = true
-
-            } label: {
-            
-                Text("scan_instructions_next", bundle: .module)
-                
-            }
-            .buttonStyle(DefaultButtonStyle())
-            
-            Button {
-                
-                retry()
-                
-            } label: {
-            
-                Text("scan_instructions_retry", bundle: .module)
-                    .font(.subheadline)
-                    .foregroundColor(configViewModel.theme.button)
-                
-            }
-            
-            NavigationLink(isActive: $showNext) {
-                
-                configViewModel.nextStepView
-                
-            } label: {
-                EmptyView()
-            }
-            .hidden()
-            
-        }
-        .frame(maxWidth: .infinity, alignment: .center)
-        
-    }
-    
-}
-
-internal struct DeniedFooter: View {
-    
-    // MARK: - Properties
-    var retry: () -> Void
-    
-    @EnvironmentObject private var configViewModel: ConfigViewModel
-    
-    // MARK: - Main
-    var body: some View {
-        
-        VStack(spacing: 30) {
-            
-            VStack(spacing: 10) {
-                
-                Text("scan_instructions_denied", bundle: .module)
-                    .font(.subheadline)
-                    .bold()
-                    .foregroundColor(configViewModel.theme.title)
-                
-                Text("scan_instructions_cannot_validated", bundle: .module)
-                    .font(.subheadline)
-                    .foregroundColor(configViewModel.theme.subtitle)
-                    .multilineTextAlignment(.center)
-                
-            }
-            
-            VStack(spacing: 20) {
-            
-                Button {
-                    
-                    retry()
-                    
-                } label: {
-                
-                    Text("scan_instructions_retry", bundle: .module)
-                    
-                }
-                .buttonStyle(DefaultButtonStyle())
-                
-                Button {
-                    
-                    dismiss()
-                    
-                } label: {
-                    
-                    Text("scan_instructions_leave", bundle: .module)
-                        .font(.subheadline)
-                        .foregroundColor(configViewModel.theme.button)
-                    
-                }
-                
-            }
-
-        }
-        .frame(maxWidth: .infinity, alignment: .center)
-        
-    }
-    
-}
-
 struct ScanInstructionsView_Previews: PreviewProvider {
     static var previews: some View {
