+++ Sources/Alloy/Public/Model/AlloySettings.swift
@@ -36,13 +36,13 @@ public class AlloySettings {
     /// Evaluate document on upload image/file
     public var evaluateOnUpload = false
     
-    /// Maximum evaluation attemps (retries(
+    /// Maximum evaluation attemps when retake images
     public var maxEvaluationAttempts: Int = 2
     
     /// Configure SDK with production environment (default is false)
     public var production = false
 
-    /// Configure SDK for make tests in environment
+    /// Configure SDK for make tests in environment (default is false)
     public var realProduction = false
 }
